requirements.txt
appnope==0.1.0
backcall==0.2.0
bcrypt==3.2.0
bitarray==1.5.3
branca==0.4.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
commonmark==0.9.1
cryptography==3.1.1
decorator==4.4.2
enum34==1.1.10
folium==0.11.0
idna==2.10
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
Jinja2==2.11.2
littletable==1.1.0
lxml==4.5.2
MarkupSafe==1.1.1
ncclient==0.6.9
numpy==1.19.2
paramiko==2.7.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.7
ptyprocess==0.6.0
pyang==2.3.2
pyangbind==0.8.1
pycparser==2.20
Pygments==2.7.1
PyNaCl==1.4.0
PyYAML==5.3.1
regex==2020.9.27
requests==2.24.0
rich==7.1.0
six==1.15.0
traitlets==5.0.4
typing-extensions==3.7.4.3
urllib3==1.25.10
wcwidth==0.2.5
xmltodict==0.12.0
appdirs==1.4.4
bcrypt==3.2.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
commonmark==0.9.1
conda==4.3.16
cryptography==3.1.1
distlib==0.3.1
filelock==3.0.12
idna==2.10
littletable==1.1.0
lxml==4.5.2
ncclient==0.6.9
paramiko==2.7.2
pipenv==2020.8.13
pycosat==0.6.3
pycparser==2.20
Pygments==2.7.1
PyNaCl==1.4.0
requests==2.24.0
rich==7.1.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.25.10
virtualenv==20.0.31
virtualenv-clone==0.5.4
xmltodict==0.12.0

